inference/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow-cpu==2.14.0
Pillow
numpy<2
